Have most pieces in place
I have most of my pieces in place, I just need to fully implement them
into my part of the code. In this commit, i added quality control (qc)
methods to qc the data according to the specs given in the assignment.
They HAVE NOT been put into use yet. That is probably my next project.
I also added references to a region’s parent region. That made the
ingest a little easier. TAKE NOTE: I had to add a “type” parameter to
readTextFile() in order to get the new switch statement to work
correctly. So that will have to be added later on. My next steps are as
follows: 1. Implement qc methods into both the binary and text file
ingest. 2. Fix add() so it adds the region to the model lists AND to
its parent region. 3. Make the program remove sub regions of a region
if said region is deleted. 4. Fix file writer to write out the correct
.txt files. 5. Document the crap out of everything so y’all understand
what the heck I’m doing