Can now Add, delete, or edit continents
The only problem so far is I can’t find a way to update names on the
continents jlist if you edit it while still being able to delete it.
Because you can’t change keys in linked hash maps. So they key has to
stay the same but the name is different in the continent object. Will
continue to look into it though. Will implement add, edit, and delete
on the other region types tomorrow.