Points of Interest maps work (sorta)
The more specific Point of Interest maps work, but it's a little
unintuitive. For example, if a Point of Interest belongs to a city in
North America, then the "Points of Interest Within City" will show it,
but not "Points of Interest Within North America".

This is going to need some kind of workaround for the Cities Within
Continents, as none of them technically "belong" to continents.